requirements.txt
beautifulsoup4==4.12.2
discord.py==2.3.2
pandas==2.1.4
python-dotenv==1.0.0
requests==2.26.0
